+++ apps/web/ui/partners/lander/lander-hero.tsx
